requirements.txt
Flask
gunicorn
requests
python-dotenv
supabase
werkzeug
